Download from Onedrive
I am seeking an alternative solution to using the costly Microsoft Azure for my needs. As a result, I am looking to develop a custom API to facilitate automatic downloads from OneDrive and enable the storage of files in specific folders within my local directory. This approach will not only offer me more cost-effective options but also empower me with the flexibility to tailor the process to my exact requirements.